refactor: move account related functions to separate files (#11870)
* refactor: move account related functions to separate files

* test: fix ut

* test: add ut
