+++ packages/vscode-extension/src/handlers/checkCopilotAccess.ts
@@ -6,7 +6,6 @@ import M365TokenInstance from "../commonlib/m365Login";
 import { signedIn } from "../commonlib/common/constant";
 import { localize } from "../utils/localizeUtils";
 import VsCodeLogInstance from "../commonlib/log";
-import { signInM365 } from "../handlers";
 import { FxError, Result, err, ok } from "@microsoft/teamsfx-api";
 import {
   AppStudioScopes,
@@ -15,6 +14,7 @@ import {
   SummaryConstant,
 } from "@microsoft/teamsfx-core";
 import { wrapError } from "../error/common";
+import { signInM365 } from "../utils/accountUtils";
 
 export async function checkCopilotAccessHandler(): Promise<Result<null, FxError>> {
   // check m365 login status, if not logged in, pop up a message
